backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
numpy
langchain-community
langchain=0.0.349
pinecone